Refactor settings and import/export components
