+++ src/components/ImportExport/index.tsx
@@ -1,20 +1,11 @@
 import React, { useState, useRef } from 'react';
-import { 
-  Download, 
-  Upload, 
-  FileText, 
-  Database, 
-  Settings,
-  CheckCircle,
-  AlertCircle,
-  X,
-  File,
-  FolderOpen,
-  Lock
-} from 'lucide-react';
-import { Connection } from '../types/connection';
-import { useConnections } from '../contexts/ConnectionContext';
-import { CollectionManager } from '../utils/collectionManager';
+import { Download, Upload, X } from 'lucide-react';
+import { Connection } from '../../types/connection';
+import { useConnections } from '../../contexts/ConnectionContext';
+import { CollectionManager } from '../../utils/collectionManager';
+import ExportTab from './ExportTab';
+import ImportTab from './ImportTab';
+import { ImportResult } from './types';
 import CryptoJS from 'crypto-js';
 
 interface ImportExportProps {
@@ -22,12 +13,6 @@ interface ImportExportProps {
   onClose: () => void;
 }
 
-interface ImportResult {
-  success: boolean;
-  imported: number;
-  errors: string[];
-  connections: Connection[];
-}
 
 export const ImportExport: React.FC<ImportExportProps> = ({ isOpen, onClose }) => {
   const { state, dispatch } = useConnections();
@@ -442,223 +427,32 @@ export const ImportExport: React.FC<ImportExportProps> = ({ isOpen, onClose }) =
 
           {/* Export Tab */}
           {activeTab === 'export' && (
-            <div className="space-y-6">
-              <div>
-                <h3 className="text-lg font-medium text-white mb-4">Export Connections</h3>
-                <p className="text-gray-400 mb-4">
-                  Export your connections to a file. Configure encryption and password options below.
-                </p>
-                
-                <div className="bg-gray-700 rounded-lg p-4 mb-4">
-                  <div className="flex items-center justify-between mb-2">
-                    <span className="text-gray-300">Total Connections:</span>
-                    <span className="text-white font-medium">{state.connections.length}</span>
-                  </div>
-                  <div className="flex items-center justify-between">
-                    <span className="text-gray-300">Groups:</span>
-                    <span className="text-white font-medium">
-                      {state.connections.filter(c => c.isGroup).length}
-                    </span>
-                  </div>
-                </div>
-              </div>
-
-              <div>
-                <label className="block text-sm font-medium text-gray-300 mb-2">
-                  Export Format
-                </label>
-                <div className="grid grid-cols-3 gap-3">
-                  {[
-                    { value: 'json', label: 'JSON', icon: FileText, desc: 'Structured data format' },
-                    { value: 'xml', label: 'XML', icon: Database, desc: 'sortOfRemoteNG compatible' },
-                    { value: 'csv', label: 'CSV', icon: Settings, desc: 'Spreadsheet format' }
-                  ].map(format => (
-                    <button
-                      key={format.value}
-                      onClick={() => setExportFormat(format.value as any)}
-                      className={`p-4 rounded-lg border-2 transition-colors ${
-                        exportFormat === format.value
-                          ? 'border-blue-500 bg-blue-500/20'
-                          : 'border-gray-600 hover:border-gray-500'
-                      }`}
-                    >
-                      <format.icon size={24} className="mx-auto mb-2 text-gray-300" />
-                      <div className="text-white font-medium">{format.label}</div>
-                      <div className="text-xs text-gray-400 mt-1">{format.desc}</div>
-                    </button>
-                  ))}
-                </div>
-              </div>
-
-              <div className="space-y-4">
-                <label className="flex items-center space-x-2">
-                  <input
-                    type="checkbox"
-                    checked={includePasswords}
-                    onChange={(e) => setIncludePasswords(e.target.checked)}
-                    className="rounded border-gray-600 bg-gray-700 text-blue-600"
-                  />
-                  <span className="text-gray-300">Include passwords in export</span>
-                </label>
-
-                <label className="flex items-center space-x-2">
-                  <input
-                    type="checkbox"
-                    checked={exportEncrypted}
-                    onChange={(e) => setExportEncrypted(e.target.checked)}
-                    className="rounded border-gray-600 bg-gray-700 text-blue-600"
-                  />
-                  <span className="text-gray-300">Encrypt export file</span>
-                  <Lock size={16} className="text-yellow-400" />
-                </label>
-
-                {exportEncrypted && (
-                  <div>
-                    <label className="block text-sm font-medium text-gray-300 mb-2">
-                      Encryption Password
-                    </label>
-                    <input
-                      type="password"
-                      value={exportPassword}
-                      onChange={(e) => setExportPassword(e.target.value)}
-                      className="w-full px-3 py-2 bg-gray-700 border border-gray-600 rounded-md text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-blue-500"
-                      placeholder="Enter encryption password"
-                    />
-                  </div>
-                )}
-              </div>
-
-              <button
-                onClick={handleExport}
-                disabled={isProcessing || state.connections.length === 0 || (exportEncrypted && !exportPassword)}
-                className="w-full py-3 bg-blue-600 hover:bg-blue-700 disabled:bg-gray-600 disabled:cursor-not-allowed text-white rounded-lg transition-colors flex items-center justify-center space-x-2"
-              >
-                {isProcessing ? (
-                  <>
-                    <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white"></div>
-                    <span>Exporting...</span>
-                  </>
-                ) : (
-                  <>
-                    <Download size={16} />
-                    <span>Export Connections</span>
-                  </>
-                )}
-              </button>
-            </div>
+            <ExportTab
+              connections={state.connections}
+              exportFormat={exportFormat}
+              setExportFormat={setExportFormat}
+              includePasswords={includePasswords}
+              setIncludePasswords={setIncludePasswords}
+              exportEncrypted={exportEncrypted}
+              setExportEncrypted={setExportEncrypted}
+              exportPassword={exportPassword}
+              setExportPassword={setExportPassword}
+              isProcessing={isProcessing}
+              handleExport={handleExport}
+            />
           )}
 
           {/* Import Tab */}
           {activeTab === 'import' && (
-            <div className="space-y-6">
-              <div>
-                <h3 className="text-lg font-medium text-white mb-4">Import Connections</h3>
-                <p className="text-gray-400 mb-4">
-                  Import connections from JSON, XML, or CSV files. Encrypted files are automatically detected.
-                </p>
-              </div>
-
-              {!importResult && (
-                <div className="border-2 border-dashed border-gray-600 rounded-lg p-8 text-center">
-                  <FolderOpen size={48} className="mx-auto mb-4 text-gray-400" />
-                  <p className="text-gray-300 mb-4">Select a file to import connections</p>
-                  <button
-                    onClick={handleImport}
-                    disabled={isProcessing}
-                    className="px-6 py-2 bg-blue-600 hover:bg-blue-700 disabled:bg-gray-600 text-white rounded-lg transition-colors flex items-center space-x-2 mx-auto"
-                  >
-                    {isProcessing ? (
-                      <>
-                        <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white"></div>
-                        <span>Processing...</span>
-                      </>
-                    ) : (
-                      <>
-                        <File size={16} />
-                        <span>Choose File</span>
-                      </>
-                    )}
-                  </button>
-                  <p className="text-xs text-gray-500 mt-2">
-                    Supported formats: .json, .xml, .csv (encrypted files supported)
-                  </p>
-                </div>
-              )}
-
-              {importResult && (
-                <div className="space-y-4">
-                  <div className={`p-4 rounded-lg border ${
-                    importResult.success 
-                      ? 'border-green-500 bg-green-500/20' 
-                      : 'border-red-500 bg-red-500/20'
-                  }`}>
-                    <div className="flex items-center space-x-2 mb-2">
-                      {importResult.success ? (
-                        <CheckCircle size={20} className="text-green-400" />
-                      ) : (
-                        <AlertCircle size={20} className="text-red-400" />
-                      )}
-                      <span className={`font-medium ${
-                        importResult.success ? 'text-green-400' : 'text-red-400'
-                      }`}>
-                        {importResult.success ? 'Import Successful' : 'Import Failed'}
-                      </span>
-                    </div>
-                    
-                    {importResult.success && (
-                      <p className="text-gray-300">
-                        Found {importResult.imported} connections ready to import.
-                      </p>
-                    )}
-                    
-                    {importResult.errors.length > 0 && (
-                      <div className="mt-2">
-                        <p className="text-red-400 text-sm font-medium">Errors:</p>
-                        <ul className="text-red-300 text-sm mt-1">
-                          {importResult.errors.map((error, index) => (
-                            <li key={index}>• {error}</li>
-                          ))}
-                        </ul>
-                      </div>
-                    )}
-                  </div>
-
-                  {importResult.success && (
-                    <div className="flex space-x-3">
-                      <button
-                        onClick={confirmImport}
-                        className="flex-1 py-2 bg-green-600 hover:bg-green-700 text-white rounded-lg transition-colors"
-                      >
-                        Import {importResult.imported} Connections
-                      </button>
-                      <button
-                        onClick={() => setImportResult(null)}
-                        className="px-4 py-2 bg-gray-600 hover:bg-gray-700 text-white rounded-lg transition-colors"
-                      >
-                        Cancel
-                      </button>
-                    </div>
-                  )}
-
-                  {!importResult.success && (
-                    <button
-                      onClick={() => setImportResult(null)}
-                      className="w-full py-2 bg-blue-600 hover:bg-blue-700 text-white rounded-lg transition-colors"
-                    >
-                      Try Again
-                    </button>
-                  )}
-                </div>
-              )}
-
-              <input
-                ref={fileInputRef}
-                type="file"
-                accept=".json,.xml,.csv,.encrypted"
-                onChange={handleFileSelect}
-                className="hidden"
-              />
-            </div>
+            <ImportTab
+              isProcessing={isProcessing}
+              handleImport={handleImport}
+              fileInputRef={fileInputRef}
+              importResult={importResult}
+              handleFileSelect={handleFileSelect}
+              confirmImport={confirmImport}
+              cancelImport={() => setImportResult(null)}
+            />
           )}
         </div>
       </div>
