+++ src/components/ImportExport/types.ts
@@ -1,0 +1,8 @@
+import { Connection } from '../../types/connection';
+
+export interface ImportResult {
+  success: boolean;
+  imported: number;
+  errors: string[];
+  connections: Connection[];
+}
